Add support for finding the right match.
When turning an Nfa into a Dfa, use ordered sets of states instead
of unordered ones; then we can be sure to find the exact match specified
by the Nfa instead of only the shortest sub-match.

This also involves several modifications of Nfa operations, to
ensure that match priority is preserved.